+++ ui/mods/com.pa.quitch.biggergw/size.js
@@ -1,26 +1,17 @@
-// Make new system sizes selectable on the size drop-down
-
-$('#game-size').append('<option value="5">VAST</option>' +
+$("#game-size").append(
+  '<option value="5">VAST</option>' +
     '<option value="6">GIGANTIC</option>' +
     '<option value="7">RIDICULOUS</option>' +
-    '<option value="8">MARATHON</option>')
+    '<option value="8">MARATHON</option>'
+);
+locUpdateDocument();
 
-// Set system size settings and add them to numberOfSystems
-
-requireGW([
-    'shared/gw_balance'
-],
-    function (balance) {
-        balance.numberOfSystems.push(
-            108, // Vast
-            144, // Gigantic
-            186, // Ridiculous
-            234 // Marathon
-        )
-        balance.galaxySizeDiffMod.push(
-            0.95,
-            0.9,
-            0.85,
-            0.8
-        )
-    })
+requireGW(["shared/gw_balance"], function (balance) {
+  balance.numberOfSystems.push(
+    108, // Vast
+    144, // Gigantic
+    186, // Ridiculous
+    234 // Marathon
+  );
+  balance.galaxySizeDiffMod.push(0.95, 0.9, 0.85, 0.8);
+});