+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v1.0 - 2020-06-25
+
+- Enabled translation support
+
 ## v0.1 - 2018-03-03
 
- - Implemented four new galaxy sizes
+- Implemented four new galaxy sizes